Split tensionWIthEcho and fastReaction
Now it seems that we have all the representative modes
