+++ static/index.html
@@ -16,9 +16,10 @@
         <option value="basic" selected="selected">Basic</option>
         <option value="randomEcho">Random echo</option>
         <option value="flicker">Flicker</option>
-        <option value="tensionAndRandomEcho">Tension and random echo</option>
-		<option value="rain">Rain</option>
-		<option value="empty">Empty</option>
+        <option value="tensionWithEcho">Tension with echo</option>
+        <option value="fastReaction">React if trigger is fast</option>
+        <option value="rain">Rain</option>
+        <option value="empty">Empty</option>
       </select>
     </div>
     <div id="select-structure--container">
