+++ web/lib/visualisations.js
@@ -1,7 +1,8 @@
 import basic from './modes/basic'
 import flicker from './modes/flicker'
 import randomEcho from './modes/randomEcho'
-import tensionAndRandomEcho from './modes/tensionAndRandomEcho'
+import tensionWithEcho from './modes/tensionWithEcho'
+import fastReaction from './modes/fastReaction'
 import rain from './modes/rain'
 
 
@@ -9,6 +10,7 @@ export default {
 	basic,
 	flicker,
 	randomEcho,
-	tensionAndRandomEcho,
+	tensionWithEcho,
+	fastReaction,
 	rain
 }
